+++ zyn/mlp.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
-import numpy as np
-
+from zyn.backend import xp as np
+from zyn.backend import fdtype
 from zyn.tensor import Tensor
 
 
@@ -10,10 +10,10 @@ class MLP:
     def __init__(self, d_model: int, d_ff: int | None = None, std: float = 0.02):
         self.d_model = d_model
         self.d_ff = d_ff if d_ff is not None else 4 * d_model
-        self.W1 = Tensor(np.random.randn(d_model, self.d_ff).astype(np.float64) * std)
-        self.b1 = Tensor(np.zeros(self.d_ff, dtype=np.float64))
-        self.W2 = Tensor(np.random.randn(self.d_ff, d_model).astype(np.float64) * std)
-        self.b2 = Tensor(np.zeros(d_model, dtype=np.float64))
+        self.W1 = Tensor(np.random.randn(d_model, self.d_ff).astype(fdtype) * std)
+        self.b1 = Tensor(np.zeros(self.d_ff, dtype=fdtype))
+        self.W2 = Tensor(np.random.randn(self.d_ff, d_model).astype(fdtype) * std)
+        self.b2 = Tensor(np.zeros(d_model, dtype=fdtype))
 
     def __call__(self, x: Tensor) -> Tensor:
         if x.shape[-1] != self.d_model:
